requirements.txt
arrow==1.4.0
asgiref==3.11.1
binaryornot==0.6.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cookiecutter==2.7.1
cryptography==46.0.6
dj-database-url==3.1.2
dj-rest-auth==7.2.0
Django==6.0.3
django-allauth==65.15.1
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-filter==25.2
django-htmx==1.27.0
django-tailwind==4.4.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
google-auth==2.49.1
google-auth-oauthlib==1.3.1
gspread==6.2.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.4
oauthlib==3.3.1
openpyxl==3.1.5
pandas==3.0.2
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
